test(mito2): move compaction tests to dedicated file
